add check-updates entry point
